Backend/app/pyproject.toml
[project]
name = "InstaShare-backend"
version = "0.1.0"
description = "InstaShare's backend: Online platform for managing documents and share its with FastAPI."
authors = [
    { name = " InstaShare's team", email = "inoid2007@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase",
    "alembic",
    "pydantic[email]>=2.11.7",
    "python-multipart>=0.0.20",
    "psycopg2>=2.9.10",
    "python-jose>=3.5.0",
    "pytest-asyncio>=1.1.0",
    "celery[librabbitmq,redis]>=5.5.3",
    "gotrue>=2.12.4",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
